Adding more details
