+++ package.json
@@ -13,12 +13,13 @@
     "@babel/cli": "^7.8.4",
     "@babel/core": "^7.8.4",
     "babel-preset-airbnb": "^4.4.0",
+    "canvas": "^2.6.1",
+    "chai": "^4.2.0",
     "eslint": "^4.14.0",
     "eslint-config-blazity": "^1.0.4",
+    "jest": "^25.1.0",
     "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "chai": "^4.2.0",
-    "jest": "^25.1.0"
+    "react-dom": "^16.2.0"
   },
   "peerDependencies": {
     "react": "^16.x.x || ^15.6.0",
