+++ src/Utils.js
@@ -8,6 +8,34 @@ const responsiveSizes = size => [
   size * 3
 ];
 
+// check webp support
+let isWebpSupportedCache = null;
+const isWebpSupported = () => {
+  if (isWebpSupportedCache !== null) {
+    return isWebpSupportedCache
+  }
+
+  const elem =
+    typeof window !== `undefined` ? window.document.createElement(`canvas`) : {};
+  if (elem.getContext && elem.getContext(`2d`)) {
+    isWebpSupportedCache =
+      elem.toDataURL(`image/webp`).indexOf(`data:image/webp`) === 0;
+    return isWebpSupportedCache
+  }
+  return false
+};
+
+const compressAndWebp = webp => `${webp ? 'output=format:webp/' : ''}compress`;
+
+const constructURL = (handle, withWebp, baseURI) => resize => transforms =>
+  [
+    baseURI,
+    resize,
+    ...transforms,
+    compressAndWebp(isWebpSupported() && withWebp),
+    handle
+  ].join('/');
+
 const srcSet = (srcBase, srcWidths, fit, transforms) =>
   srcWidths
     .map(
@@ -28,4 +56,5 @@ const getWidths = (width, maxWidth) => {
 export {
   srcSet,
   getWidths,
+  constructURL,
 }
